Move uyuni common libs to python folder
