+++ python/uyuni/Makefile.defs
