+++ python/uyuni/common/__init__.py
