+++ python/uyuni/common/checksum.py
