+++ python/uyuni/common/cli.py
