+++ python/uyuni/common/context_managers.py
