+++ python/uyuni/common/rhn_mpm.py
