+++ python/uyuni/common/rhn_pkg.py
