+++ python/uyuni/common/rhn_rpm.py
