+++ python/uyuni/common/usix.py
