+++ python/uyuni/pylintrc
