+++ rel-eng/packages/uyuni-common-libs
@@ -1,1 +1,1 @@
-4.3.2-1 uyuni/common-libs/
+4.3.2-1 python/uyuni/
